requirements.txt
streamlit
Image
python-dotenv
google-generativeai